proc/wpc: more updates to solenoid driving code; automatically enable/disable flippers
On certain games the top 3 GI bits map to bits in core_getAllSol() that we can use to detect GameOver/GameStarted and flippers Enabled/Disabled.  At this point, it isn't entirely clear which games it applies to, so to prevent disabling flippers on unsupported games, we use the "GameOver" bit (28) to disable the flippers when the game starts up, and then make use of bit 30 to enable and disable the flippers over the course of a game.